+++ levels/level_1.lua
@@ -1,58 +1,11 @@
 local map = {}
 
-map = { etages = 4, lignes = 8, colonnes = 12,
+map = { etages = 1, lignes = 1, colonnes = 6, 
 
 
   {-- etage 1
-  {    1,    1,  1,  1,  1,  1,  1,  1,  1,  1,    1,   1 },-- Ligne 1 - 12 colonnes
-  {  "_",  "_",  1,  1,  1,  1,  1,  1,  1,  1,  "_",  "_" },-- Ligne 2 - 12 colonnes
-  {  "_",    1,  1,  1,  1,  1,  1,  1,  1,  1,    1,  "_" },-- Ligne 3- 12 colonnes
-  {    1,    1,  1,  1,  1,  1,  1,  1,  1,  1,    1,   1 },-- Ligne 4 - 12 colonnes
-  {    1,    1,  1,  1,  1,  1,  1,  1,  1,  1,    1,   1 },-- Ligne 5 - 12 colonnes
-  {  "_",    1,  1,  1,  1,  1,  1,  1,  1,  1,    1,  "_" },-- Ligne 6 - 12 colonnes
-  {  "_",  "_",  1,  1,  1,  1,  1,  1,  1,  1,  "_",  "_" },-- Ligne 7 - 12 colonnes
-  {    1,    1,  1,  1,  1,  1,  1,  1,  1,  1,    1,   1 }-- Ligne 8 - 12 colonnes
-  }
+    { 1,  1,  1,  1,  1,  1 }
 
-  ,
-
-  {-- etage 2
-  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 1 - 12 colonnes
-  {  "_",  "_",  "_",    2,    2,    2,    2,    2,    2,  "_",  "_",  "_" },-- Ligne 2 - 12 colonnes
-  {  "_",  "_",  "_",    2,    2,    2,    2,    2,    2,  "_",  "_",  "_" },-- Ligne 3- 12 colonnes
-  {  "_",  "_",  "_",    2,    2,    2,    2,    2,    2,  "_",  "_",  "_" },-- Ligne 4 - 12 colonnes
-  {  "_",  "_",  "_",    2,    2,    2,    2,    2,    2,  "_",  "_",  "_" },-- Ligne 5 - 12 colonnes
-  {  "_",  "_",  "_",    2,    2,    2,    2,    2,    2,  "_",  "_",  "_" },-- Ligne 6 - 12 colonnes
-  {  "_",  "_",  "_",    2,    2,    2,    2,    2,    2,  "_",  "_",  "_" },-- Ligne 7 - 12 colonnes
-  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" }-- Ligne 8 - 12 colonnes
-  }
-
-  ,
-
-
-  {-- etage 3
-  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 1 - 12 colonnes
-  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 2 - 12 colonnes
-  {  "_",  "_",  "_",  "_",    3,  3,  3,  3,  "_",  "_",  "_",  "_" },-- Ligne 3- 12 colonnes
-  {  "_",  "_",  "_",  "_",    3,  3,  3,  3,  "_",  "_",  "_",  "_" },-- Ligne 4 - 12 colonnes
-  {  "_",  "_",  "_",  "_",    3,  3,  3,  3,  "_",  "_",  "_",  "_" },-- Ligne 5 - 12 colonnes
-  {  "_",  "_",  "_",  "_",    3,  3,  3,  3,  "_",  "_",  "_",  "_" },-- Ligne 6 - 12 colonnes
-  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 7 - 12 colonnes
-  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" }-- Ligne 8 - 12 colonnes
-  }
-
-  ,
-
-
-  {-- etage 4
-  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 1 - 12 colonnes
-  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 2 - 12 colonnes
-  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 3- 12 colonnes
-  {  "_",  "_",  "_",  "_",  "_",  4,  4,  "_",  "_",  "_",  "_",  "_" },-- Ligne 4 - 12 colonnes
-  {  "_",  "_",  "_",  "_",  "_",  4,  4,  "_",  "_",  "_",  "_",  "_" },-- Ligne 5 - 12 colonnes
-  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 6 - 12 colonnes
-  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 7 - 12 colonnes
-  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" }-- Ligne 8 - 12 colonnes
   }
 
 }
