+++ levels/level_10.lua
@@ -1,0 +1,60 @@
+local map = {}
+
+map = { etages = 4, lignes = 8, colonnes = 12,
+
+
+  {-- etage 1
+  {    1,    1,  1,  1,  1,  1,  1,  1,  1,  1,    1,   1 },-- Ligne 1 - 12 colonnes
+  {  "_",  "_",  1,  1,  1,  1,  1,  1,  1,  1,  "_",  "_" },-- Ligne 2 - 12 colonnes
+  {  "_",    1,  1,  1,  1,  1,  1,  1,  1,  1,    1,  "_" },-- Ligne 3- 12 colonnes
+  {    1,    1,  1,  1,  1,  1,  1,  1,  1,  1,    1,   1 },-- Ligne 4 - 12 colonnes
+  {    1,    1,  1,  1,  1,  1,  1,  1,  1,  1,    1,   1 },-- Ligne 5 - 12 colonnes
+  {  "_",    1,  1,  1,  1,  1,  1,  1,  1,  1,    1,  "_" },-- Ligne 6 - 12 colonnes
+  {  "_",  "_",  1,  1,  1,  1,  1,  1,  1,  1,  "_",  "_" },-- Ligne 7 - 12 colonnes
+  {    1,    1,  1,  1,  1,  1,  1,  1,  1,  1,    1,   1 }-- Ligne 8 - 12 colonnes
+  }
+
+  ,
+
+  {-- etage 2
+  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 1 - 12 colonnes
+  {  "_",  "_",  "_",    2,    2,    2,    2,    2,    2,  "_",  "_",  "_" },-- Ligne 2 - 12 colonnes
+  {  "_",  "_",  "_",    2,    2,    2,    2,    2,    2,  "_",  "_",  "_" },-- Ligne 3- 12 colonnes
+  {  "_",  "_",  "_",    2,    2,    2,    2,    2,    2,  "_",  "_",  "_" },-- Ligne 4 - 12 colonnes
+  {  "_",  "_",  "_",    2,    2,    2,    2,    2,    2,  "_",  "_",  "_" },-- Ligne 5 - 12 colonnes
+  {  "_",  "_",  "_",    2,    2,    2,    2,    2,    2,  "_",  "_",  "_" },-- Ligne 6 - 12 colonnes
+  {  "_",  "_",  "_",    2,    2,    2,    2,    2,    2,  "_",  "_",  "_" },-- Ligne 7 - 12 colonnes
+  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" }-- Ligne 8 - 12 colonnes
+  }
+
+  ,
+
+
+  {-- etage 3
+  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 1 - 12 colonnes
+  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 2 - 12 colonnes
+  {  "_",  "_",  "_",  "_",    3,  3,  3,  3,  "_",  "_",  "_",  "_" },-- Ligne 3- 12 colonnes
+  {  "_",  "_",  "_",  "_",    3,  3,  3,  3,  "_",  "_",  "_",  "_" },-- Ligne 4 - 12 colonnes
+  {  "_",  "_",  "_",  "_",    3,  3,  3,  3,  "_",  "_",  "_",  "_" },-- Ligne 5 - 12 colonnes
+  {  "_",  "_",  "_",  "_",    3,  3,  3,  3,  "_",  "_",  "_",  "_" },-- Ligne 6 - 12 colonnes
+  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 7 - 12 colonnes
+  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" }-- Ligne 8 - 12 colonnes
+  }
+
+  ,
+
+
+  {-- etage 4
+  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 1 - 12 colonnes
+  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 2 - 12 colonnes
+  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 3- 12 colonnes
+  {  "_",  "_",  "_",  "_",  "_",  4,  4,  "_",  "_",  "_",  "_",  "_" },-- Ligne 4 - 12 colonnes
+  {  "_",  "_",  "_",  "_",  "_",  4,  4,  "_",  "_",  "_",  "_",  "_" },-- Ligne 5 - 12 colonnes
+  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 6 - 12 colonnes
+  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" },-- Ligne 7 - 12 colonnes
+  {  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_",  "_" }-- Ligne 8 - 12 colonnes
+  }
+
+}
+
+return map
